[Style] [GGFE-195] 코인내역 모달 pagination위치 수정
