+++ components/modal/store/CoinHistoryContainer.tsx
@@ -1,61 +1,26 @@
-import { useEffect, useState } from 'react';
-import { useSetRecoilState } from 'recoil';
-import { ICoinHistoryTable } from 'types/userTypes';
-import { mockInstance } from 'utils/mockAxios';
-import { errorState } from 'utils/recoil/error';
+import { ICoinHistory } from 'types/userTypes';
 import CoinHistoryDetails from 'components/modal/store/CoinHistoryDetails';
-import PageNation from 'components/Pagination';
 import styles from 'styles/modal/store/CoinHistoryContainer.module.scss';
 
-export default function CoinHistoryContainer() {
-  const [currentPage, setCurrentPage] = useState<number>(1);
-  const [coinHistoryList, setCoinHistoryList] = useState<ICoinHistoryTable>({
-    useCoinList: [],
-    totalPage: 0,
-    currentPage: 0,
-  });
-  const setError = useSetRecoilState(errorState);
+type CoinHistoryProps = {
+  useCoinList: ICoinHistory[];
+};
 
-  useEffect(() => {
-    getCoinHistoryList();
-  }, [currentPage]);
-
-  const getCoinHistoryList = async () => {
-    try {
-      const res = await mockInstance.get(
-        `/users/coin/?page=${currentPage}&size=5`
-      );
-      setCoinHistoryList({
-        useCoinList: res.data.useCoinList,
-        totalPage: res.data.totalPage,
-        currentPage: currentPage,
-      });
-    } catch (e) {
-      setError('HB03');
-    }
-  };
-
+export default function CoinHistoryContainer({
+  useCoinList,
+}: CoinHistoryProps) {
   return (
     <div className={styles.container}>
-      {coinHistoryList.useCoinList.length === 0 ? (
+      {useCoinList.length === 0 ? (
         <div className={styles.empty}>GG코인 내역이 존재하지 않습니다.</div>
       ) : (
-        coinHistoryList.useCoinList.map((coinHistory) => (
+        useCoinList.map((coinHistory) => (
           <CoinHistoryDetails
             key={coinHistory.createdAt.toString()}
             details={coinHistory}
           />
         ))
       )}
-      <div>
-        <PageNation
-          curPage={currentPage}
-          totalPages={coinHistoryList.totalPage}
-          pageChangeHandler={(pageNumber: number) => {
-            setCurrentPage(pageNumber);
-          }}
-        />
-      </div>
     </div>
   );
 }
