+++ components/modal/store/UserCoinHistoryModal.tsx
@@ -1,22 +1,52 @@
-import React from 'react';
+import { useEffect, useState } from 'react';
 import { useSetRecoilState } from 'recoil';
 import { FaCoins } from 'react-icons/fa';
-import { ICoin } from 'types/userTypes';
+import { ICoin, ICoinHistoryTable } from 'types/userTypes';
+import { mockInstance } from 'utils/mockAxios';
+import { errorState } from 'utils/recoil/error';
 import { modalState } from 'utils/recoil/modal';
 import {
   ModalButtonContainer,
   ModalButton,
 } from 'components/modal/ModalButton';
 import CoinHistoryContainer from 'components/modal/store/CoinHistoryContainer';
+import PageNation from 'components/Pagination';
 import styles from 'styles/modal/store/UserCoinHistoryModal.module.scss';
 
 export default function UserCoinHistoryModal({ coin }: ICoin) {
   const setModal = useSetRecoilState(modalState);
+  const [currentPage, setCurrentPage] = useState<number>(1);
+  const [coinHistoryList, setCoinHistoryList] = useState<ICoinHistoryTable>({
+    useCoinList: [],
+    totalPage: 0,
+    currentPage: 0,
+  });
+  const setError = useSetRecoilState(errorState);
 
   const closeModal = () => {
     setModal({
       modalName: null,
     });
+  };
+
+  useEffect(() => {
+    getCoinHistoryList();
+  }, [currentPage]);
+
+  // TODO: instance로 수정
+  const getCoinHistoryList = async () => {
+    try {
+      const res = await mockInstance.get(
+        `/users/coin/?page=${currentPage}&size=5`
+      );
+      setCoinHistoryList({
+        useCoinList: res.data.useCoinList,
+        totalPage: res.data.totalPage,
+        currentPage: currentPage,
+      });
+    } catch (e) {
+      setError('HB03');
+    }
   };
 
   return (
@@ -29,7 +59,16 @@ export default function UserCoinHistoryModal({ coin }: ICoin) {
           <FaCoins />
         </div>
       </div>
-      <CoinHistoryContainer />
+      <CoinHistoryContainer useCoinList={coinHistoryList.useCoinList} />
+      <div>
+        <PageNation
+          curPage={currentPage}
+          totalPages={coinHistoryList.totalPage}
+          pageChangeHandler={(pageNumber: number) => {
+            setCurrentPage(pageNumber);
+          }}
+        />
+      </div>
       <ModalButtonContainer>
         <ModalButton style='positive' value='확인' onClick={closeModal} />
       </ModalButtonContainer>
