+++ styles/modal/store/CoinHistoryContainer.module.scss
@@ -3,10 +3,9 @@
 .container {
   display: flex;
   width: 85%;
-  height: 50vh;
+  height: 30vh;
   flex-direction: column;
   padding-top: 1rem;
-  margin-bottom: 2rem;
   overflow-y: scroll;
   border-top: 1px solid #3b363b36;
   .empty {
